fix(runtime): avoid HTTP self-fetch in SSR; load public data via fs with Node runtime; keep fetch as fallback
